Stuff I was working on last year
This is probably all embarrassingly bad which is why I never committed it then, but I don't remember writing it any more so the sting of however embarrassing it might be has passed. Maybe it's a window into how I think about breaking down problems, even if it is not all that polished or good for anything.